+++ src/app/shell/personal-cabinet/parent/create-child/child-form/child-form.component.scss
