+++ src/app/shell/personal-cabinet/parent/create-child/create-child.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { ParentCreateChildComponent } from './parent-create-child.component';
+import { CreateChildComponent } from './create-child.component';
 import { NgxsModule, Store } from '@ngxs/store';
 import { MockStore } from '../../../../shared/mocks/mock-services';
 import { FormGroup, FormsModule, ReactiveFormsModule } from '@angular/forms';
@@ -7,9 +7,9 @@ import { MatIconModule } from '@angular/material/icon';
 import { Component, Input } from '@angular/core';
 import { RouterTestingModule } from '@angular/router/testing';
 
-describe('ParentCreateChildComponent', () => {
-  let component: ParentCreateChildComponent;
-  let fixture: ComponentFixture<ParentCreateChildComponent>;
+describe('CreateChildComponent', () => {
+  let component: CreateChildComponent;
+  let fixture: ComponentFixture<CreateChildComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
@@ -24,14 +24,14 @@ describe('ParentCreateChildComponent', () => {
         { provide: Store, useValue: MockStore },
       ],
       declarations: [
-        ParentCreateChildComponent,
+        CreateChildComponent,
         MockChildFormComponent]
     })
-    .compileComponents();
+      .compileComponents();
   });
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(ParentCreateChildComponent);
+    fixture = TestBed.createComponent(CreateChildComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
