+++ src/app/shell/personal-cabinet/parent/parent.module.ts
@@ -1,8 +1,7 @@
-import { APP_INITIALIZER, NgModule } from '@angular/core';
+import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ParentRoutingModule } from './parent-routing.module';
 import { ParentConfigComponent } from './parent-config/parent-config.component';
-import { MatButtonToggleModule } from '@angular/material/button-toggle';
 import { MatCardModule } from '@angular/material/card';
 import { MatButtonModule } from '@angular/material/button';
 import { MatChipsModule } from '@angular/material/chips';
@@ -12,9 +11,9 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 import { MatRadioModule } from '@angular/material/radio';
 import { MatSelectModule } from '@angular/material/select';
 import { MatIconModule } from '@angular/material/icon';
-import { ChildFormComponent } from './parent-create-child/child-form/child-form.component'
+import { ChildFormComponent } from './create-child/child-form/child-form.component'
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { ParentCreateChildComponent } from './parent-create-child/parent-create-child.component';
+import { CreateChildComponent } from './create-child/create-child.component';
 import { ParentInfoComponent } from './parent-info/parent-info.component';
 import { CreateApplicationComponent } from '../../workshop-details/create-application/create-application.component';
 import { SharedModule } from 'src/app/shared/shared.module';
@@ -24,7 +23,7 @@ import { MatCheckboxModule } from '@angular/material/checkbox';
 @NgModule({
   declarations: [
     ParentConfigComponent,
-    ParentCreateChildComponent,
+    CreateChildComponent,
     ChildFormComponent,
     ParentInfoComponent,
     CreateApplicationComponent
